Use annotation to name arguments
Removes this old need to compile it with some weird debugging parameters.
Backstory: Compiled java code discards the names of variables, so the old way of reading this used to produce arg0, arg1, arg2 etc if the file wasn't compiled with various debug settings.  Instead add annotation to command args with a 'name' field and specify this as a Runtime annotation (which it already is)