Provide an optional console input/output for running bot command (#11)
* Run with --console to support a console command interface
I.e. you can type bot commands in on the command line and get responses there.
Requires your bot to have the CnC module loaded or it'll just break.
Made optional just in case someone relies on the current behaviour.

* Not sure why this is necessary and don't like it.
One day I'll have to rewrite this class, for now we assume LL wont mint any untrustworthy certs that we as a SL bot care about (...)
I.e. the risk as a stand alone SL centric project is small, however this pollutes the current JVM entirely so as a module this behaviour is undesirable.
Much sighing.